Styling fixes and move components
