+++ client/src/components/Gallery/DiscussionBoard.less
