+++ client/src/components/Gallery/FilterComponent.jsx
