+++ client/src/components/Gallery/GalleryObjectForm.less
@@ -4,10 +4,6 @@ div.GalleryObjectForm {
   width: 80%;
   margin: 0.5rem;
   padding: 0 2rem;
-  &:hover {
-    cursor: pointer;
-    transform: scale(1.05);
-  }
 
   .ant-form {
     max-height: 50%;
@@ -17,7 +13,7 @@ div.GalleryObjectForm {
     border-radius: inherit;
     background-color: #F4F4F5;
     position: relative;
-    
+
     .ant-form-item {
       display: block;
       width: 100%;
@@ -25,12 +21,13 @@ div.GalleryObjectForm {
       .ant-select {
         width: 100%;
       }
-            }
+    }
 
-      /* Style for the input element (text box) */
-      input, textarea {
-        width: 80%; 
-      
+    /* Style for the input element (text box) */
+    input,
+    textarea {
+      width: 80%;
+
     }
   }
 }
@@ -41,7 +38,8 @@ div.GalleryObjectFormWrapper.visible {
   border-radius: 15px;
   position: relative;
   background-color: #F4F4F5;
-  transition: all 0.5s ease; /* Add transition effect */
+  transition: all 0.5s ease;
+  /* Add transition effect */
   width: 25%;
 }
 
@@ -52,8 +50,10 @@ div.GalleryObjectFormWrapper {
   border-radius: 15px;
   position: relative;
   background-color: #F4F4F5;
-  transition: all 0.5s ease; /* Add transition effect */
-  height: 50%; 
+  transition: all 0.5s ease;
+  /* Add transition effect */
+  height: 50%;
   width: 11%;
-  overflow: hidden; /* Hide overflow content */
+  overflow: hidden;
+  /* Hide overflow content */
 }
