+++ client/src/components/Gallery/Search.jsx
