+++ client/src/components/Gallery/Search.less
