+++ client/src/components/Gallery/Share.jsx
