+++ client/src/components/Gallery/thumbnail.png
