+++ client/src/components/Gallery/thumbnails.png
