+++ client/src/views/BlocklyPage/BlocklyPage.jsx
@@ -9,7 +9,7 @@ import {
   getActivityToolboxAll,
 } from "../../Utils/requests"
 import { useGlobalState } from "../../Utils/userState"
-import  GalleryObjectForm  from "../Gallery/GalleryObjectForm"
+import GalleryObjectForm from "../../components/Gallery/GalleryObjectForm"
 
 //TO SUBMIT TO GALLERY
