+++ client/src/views/Gallery/Gallery.less
@@ -28,7 +28,7 @@ div.galleryItem {
     &>div.flex-row {
         max-width: 100%;
         break-inside: avoid;
-        justify-content: space-between;		
+        justify-content: space-between;
     }
 }
 
@@ -60,4 +60,5 @@ p {
 
 div.galleryRows {
     margin: 1.5rem 0 0 0;
+    max-height: 60dvh;
 }
